fixed dht22 driver and example

+++ examples/platform-specific/zoul/test-dht22.c
@@ -55,7 +55,7 @@ static struct etimer et;
 /*---------------------------------------------------------------------------*/
 PROCESS_THREAD(remote_dht22_process, ev, data)
 {
-  int16_t temperature, humidity;
+  int temperature, humidity;
 
   PROCESS_BEGIN();
   SENSORS_ACTIVATE(dht22);
